+++ changelog.md
@@ -1,1 +1,2 @@
-backported a fix
+- fixed effect providers recipes being broken
+- neutral effect will be correctly marked as such